requirements.txt
fastapi 
uvicorn
pydantic[email]
ppsycopg2 
reverse_geocoder